Add Builder.customViewRaw and cleanup padding
customViewRaw allows bypassing the default scrollview and padding applied to
custom views, for example if the custom view has it's own scrollable content.

This also makes minor changes to how padding is distributed, but changes to the
implementation are major.

The title bar is now always fixed (previously it scrolled for custom content)
The padding between the title bar and content is now split evenly between
the title bar topMargin and the content paddingTop, if the content is scrollable
the divider is placed half way between.

Likewise the button bar padding is split between the button bar and content.