Refactor padding to be more efficient
The previous setup was causing lag when showing or updating the dialog,
because it needed to wait until everything was laid out before adjusting
the padding, which would trigger a second layout.

This method uses a custom ViewGroup (MDRootLayout) which determines
padding and button stacking onMeasure.

Changes were also made with how the padding was split up, to avoid having to
move padding between views.