Use sensible defaults for widths on tablets.
This implements some sensible widths for dialogs on tablets. Before, they often went nearly edge-to-edge with some minor padding on the sides, and the material design spec doesn't provide much help here. We had a long discussion about how best to go about this, with the ideal solution being to automagically create the best experience based on preferred line length. Handling every possible case of that is going to be a lot of tedious work though.

This is a simpler approach, where we basically say up front what we believe a good width is. More specifically, I stole these widths from AOSP's spec for what the notification panel width should be. Toyed with them in another library, and it turns out they look pretty good here. This is by far the simplest approach, and covers probably 98% of all the use cases.