Suggestion about core version
I updated older Material-Dialogs version to newer one, but I failed to import core:0.8.6.0 version. I can't see the tag/release of 0.8.6.0 version and I assume that that is a reason of this.

So I suggest that core:0.8.5.2 version should be in the README file because I succeeded to import this.