Allow Color, File & Folder Dialogs to be hosted in Fragments
The ColorDialog, FileChooserDialog & FolderChooserDialog can now be
hosted in a Fragment (rather than just an Activity) and the Fragment
itself can implement the Dialog's callbacks.

Note: The `Dialog.show()` method now requires the FragmentManager to be
passed in from the caller, as we can no longer cast the Context to an
Activity to get a reference to the FragmentManager.