Add overloads for non-default behavior of Color/Folder/File dialogs
Note: This is still a signature change for `show()`, we have to pass in either a FragmentActivity or FragmentManager. This is because it’s impossible for the dialog builder to determine whether the fragment was shown from an Activity or Fragment, so it can’t glean the fragment manager from the passed in context.